+++ packages/e2e/tests/02-organization/02-teams/02-02-03-02-team-members-add-created.test.ts
@@ -7,7 +7,7 @@ import { waitForTeamRow } from '@/test-helpers/team-helpers';
 import { createUserViaUI } from '@/test-helpers/user-helpers';
 import type { Page } from '@playwright/test';
 
-test.describe('User Team Assignment Tests', () => {
+test.describe('User Team Assignment Tests - Add Created User', () => {
   test.describe.configure({ timeout: 60000 });
   let dashboardPage: DashboardPage;
   let loginPage: LoginPage;
@@ -234,67 +234,6 @@ test.describe('User Team Assignment Tests', () => {
     const userInList = membersList
       .locator('.ant-list-item-meta-title')
       .filter({ hasText: userEmail });
-    await expect(userInList).toBeVisible({ timeout: 5000 });
-
-    testReporter.completeStep('Verify user in team members', 'passed');
-
-    await testReporter.finalizeTest();
-  });
-
-  test('should add specific user to team by email @system @users @teams', async ({
-    page,
-    screenshotManager: _screenshotManager,
-    testReporter,
-    testDataManager,
-  }) => {
-    const userToAdd = await createUserViaUI(page, testDataManager);
-    await dismissCreateUserModal(page);
-    const teamName = 'Private Team';
-
-    await ensureUserActive(page, userToAdd.email);
-
-    testReporter.startStep('Navigate to Teams section');
-
-    await dismissCreateUserModal(page);
-    const nav = new NavigationHelper(page);
-    await nav.goToOrganizationTeams();
-    await waitForTeamsPage(page);
-    await fillResourceSearch(page, teamName);
-
-    testReporter.completeStep('Navigate to Teams section', 'passed');
-
-    testReporter.startStep(`Add user ${userToAdd.email} to team`);
-
-    await openTeamMembersDialog(page, teamName);
-
-    const teamModal = page.locator('.ant-modal').filter({ hasText: 'Manage Team Members' });
-    await expect(teamModal).toBeVisible({ timeout: 5000 });
-
-    const addMemberTab = page.getByRole('tab', { name: 'Add Member' });
-    await expect(addMemberTab).toBeVisible();
-    await addMemberTab.click();
-
-    await selectUserInAddMemberTab(page, userToAdd.email);
-
-    // Wait for API response to complete
-    await page.waitForLoadState('networkidle');
-
-    testReporter.completeStep(`Add user ${userToAdd.email} to team`, 'passed');
-
-    testReporter.startStep('Verify user in team members');
-
-    await expect(teamModal).toBeVisible({ timeout: 5000 });
-
-    const currentMembersTab = page.getByRole('tab', { name: 'Current Members' });
-    await expect(currentMembersTab).toBeVisible();
-    await currentMembersTab.click();
-
-    const membersList = teamModal.locator('.ant-list-items');
-    await expect(membersList).toBeVisible();
-
-    const userInList = membersList
-      .locator('.ant-list-item-meta-title')
-      .filter({ hasText: userToAdd.email });
     await expect(userInList).toBeVisible({ timeout: 5000 });
 
     testReporter.completeStep('Verify user in team members', 'passed');
